Chainlit/chain_hello/requirements.txt
chainlit
python-dotenv
openai
speech_recognition
pyaudio 